- Fixed list of errors in the documentation to match implementation: WSA* errors are always returned. - Fixed race in SCTPDispatchConnectRequest: the connect call would appear to succeed because the entire   process took place before soconnect returned. - The driver package is no longer listed in Add/Remove Programs as a separate entry on Vista and later. - Removed echo_server3 from the build since it's not generally useful.

+++ apps/inc/ws2sctp.h
@@ -405,8 +405,8 @@ struct sctp_event_subscribe {
  * ancillary data structures
  */
 struct sctp_initmsg {
-	unsigned long sinit_num_ostreams;
-	unsigned long sinit_max_instreams;
+	unsigned short sinit_num_ostreams;
+	unsigned short sinit_max_instreams;
 	unsigned short sinit_max_attempts;
 	unsigned short sinit_max_init_timeo;
 };
